refactor TaskService

+++ backend/src/main/java/com/codecool/quizzzz/controller/TaskController.java
@@ -21,7 +21,7 @@ public class TaskController {
   }
 
   @GetMapping("/quiz/detailed/{quizId}")
-  public ResponseEntity<List<DetailedTaskDTO>> getAllDetailedTasksByQuiz(@PathVariable int quizId) {
+  public ResponseEntity<List<DetailedTaskDTO>> getAllDetailedTasksByQuiz(@PathVariable Long quizId) {
     return ResponseEntity.ok().body(taskService.getAllDetailedByQuiz(quizId));
   }
 
@@ -41,22 +41,22 @@ public class TaskController {
   }
 
   @PatchMapping("/{taskId}")
-  public ResponseEntity<Integer> updateTask(@PathVariable int taskId, @RequestBody QuestionDTO questionDTO) {
+  public ResponseEntity<Long> updateTask(@PathVariable Long taskId, @RequestBody QuestionDTO questionDTO) {
     return ResponseEntity.ok().body(taskService.update(taskId, questionDTO));
   }
 
   @GetMapping("/quiz/{quizId}/{taskIndex}")
-  public ResponseEntity<TaskDTO> getTask(@PathVariable int quizId, @PathVariable int taskIndex) {
+  public ResponseEntity<TaskDTO> getTask(@PathVariable Long quizId, @PathVariable int taskIndex) {
     return ResponseEntity.ok().body(taskService.getTask(quizId, taskIndex));
   }
 
   @GetMapping("/{taskId}")
-  public ResponseEntity<TaskDTO> getTask(@PathVariable int taskId) {
+  public ResponseEntity<TaskDTO> getTask(@PathVariable Long taskId) {
     return ResponseEntity.ok().body(taskService.getTask(taskId));
   }
 
   @DeleteMapping("/{taskId}")
-  public ResponseEntity<Boolean> deleteTask(@PathVariable int taskId) {
+  public ResponseEntity<Boolean> deleteTask(@PathVariable Long taskId) {
     return ResponseEntity.ok().body(taskService.deleteTask(taskId));
   }
 }
