+++ backend/src/main/java/com/codecool/quizzzz/dto/task/DetailedTaskDTO.java
@@ -4,6 +4,6 @@ import com.codecool.quizzzz.dto.answer.DetailedAnswerDTO;
 
 import java.util.List;
 
-public record DetailedTaskDTO(int taskId, int quizId, int taskIndex, String questionText,
+public record DetailedTaskDTO(Long taskId, Long quizId, int taskIndex, String questionText,
                               List<DetailedAnswerDTO> answers) {
 }
